Put tests in tests folder
